Add Generators
Adds a Rails generator for installing everything needed to get up and running.
The generator includes a "Hello World" example that allows the user to see an
example of working code. All examples represent our current best practices
at Shaka Code.

The generator is invoked as follows:

`rails generate react_on_rails:install`

The generator can also be passed several options:

- `--redux` (alias: `-R`): uses Redux in the Hello World example client code
- `--server-rendering` (alias: `S`): include files needed for server rendering

These options can be combined:

`rails generate react_on_rails:install --redux --server-rendering`

All generator code is inside of lib/generators/react_on_rails

All generator specs are included inside of the spec/react_on_rails/generators
folder. These tests will create a dummy folder specifically for testing the generators.
Any files that the generator expects to already be existing when run will need to be
manually created as part of the setup phase for each test.

The install generator relies on several other react_on_rails generators that are
hidden from the user via the special `!hide` command at the beginning of the
generator's definition. These generators are invoked from the install generator
according to the options passed from the user.

Because of the multiple unique combinations of options that can be passed from
the user, it was necessary to split template files into several sub-folders based
on those options. Files common to all builds are inside of the templates/base
folder and files specific to non-redux or redux buildes are in similarly-named
folders. Within each, there may or not be further folders depending
on whether or not the user has chosen to also pass server-rendering as an option.

This commit was the work of several developers:

- Aaron Van Bokhoven (original committer)
- Rob Wise (generator code, enhancments, tests, revisions, documentation)
- Blaine Hatab (bug fixes and revisions)
- Justin Gordon (extensive code reviews, revisions, documentation)

List of partial updates during development of generators:

- Add link to blog post in README
- Implement Redux generation
- Put screenshot of client-side versus server-side react component in README
- Change name of Generator Testing Script to generator_testing_script.md
- do not include other hello world example along with Redux hello world example
- Make 'basic' example demonstrate state'
- Make naming more consistent across different versions of generation
- added prerender false if it isnt built with server rendering
- prevent user from running genertor with uncommitted changes
- Convert no-redux components to es2015 class syntax
- Use 'static' keyword for displayName in redux components
- Avoid using constructor to set initial state in non-redux generator
- Add trailing comma to base webpack config
- added check for development to reload server bundle ever time it changes. changed links in readme from shell to markdown links. moved linters order
- Add heroku_development_generator
- Create a Rails Sass helper file
- Rename pre and post bootstrap partials and improve comments
- Switch to Alex's suggested React syntax and add client readme
- Force git to use capital filenames where appropriate
- Make exception for eslinter for e and _ variables
- Add comment about differing production and dev builds to middleware
- Change Redux alias to -R from -d
- Rearrange README section ordering as directed by JG
- Add styling to post-bootstrap to prove it's working